STSP.jl

STSP.jl

Index

Index

STSP.AbstractCompConnexe

STSP.AbstractEdge

STSP.AbstractGraph

STSP.AbstractNode

STSP.AbstractPriorityItem

STSP.AbstractQueue

STSP.Edge

STSP.Graph

STSP.Node

STSP.PriorityItem

STSP.PriorityQueue

STSP.Queue

Base.:==

Base.:==

Base.in

Base.isless

Base.length

Base.popfirst!

Base.popfirst!

Base.push!

Base.show

Base.show

Base.show

Base.show

STSP.Algorithme_HK

STSP.Algorithme_Prim

STSP.Algortihme_Kruskal

STSP.add_edge!

STSP.add_node!

STSP.add_symmetry!

STSP.affichage_RSL

STSP.affichage_arbre_minimal_kruskal

STSP.affichage_arbre_minimal_prim

STSP.complete_graph_edges

STSP.data

STSP.data

STSP.decalage

STSP.edges

STSP.get_degree

STSP.get_one_tree

STSP.increment_edges

STSP.increment_nodes

STSP.is_empty

STSP.n_nodes_to_read

STSP.name

STSP.name

STSP.name

STSP.nb_edges

STSP.nb_nodes

STSP.nodes

STSP.noeud_1

STSP.noeud_2

STSP.one_tree!

STSP.parcours_preordre

STSP.plot_graph

STSP.plot_graph

STSP.read_edges

STSP.read_header

STSP.read_nodes

STSP.read_stsp

STSP.remove_edge!

STSP.remove_node!

STSP.symetric

STSP.AbstractCompConnexe — Type

Type abstrait dont d'autres types de composantes connexes dériveront.

STSP.AbstractEdge — Type

Type abstrait dont d'autres types d'arêtes dériveront.

STSP.AbstractGraph — Type

Type abstrait dont d'autres types de graphes dériveront.

STSP.AbstractNode — Type

Type abstrait dont d'autres types de noeuds dériveront.

STSP.AbstractPriorityItem — Type

Type abstrait pour les éléments de priorité.

STSP.AbstractQueue — Type

Type abstrait dont d'autres types de files dériveront.

STSP.Edge — Type

Type représentant les arêtes d'un graphe.

Exemple:

    edge = Edge("James", 10, node1, node2)

STSP.Graph — Type

Type representant un graphe comme un ensemble de noeuds.

Exemple :

node1 = Node("Joe", 3.14)
node2 = Node("Steve", exp(1))
node3 = Node("Jill", 4.12)
G = Graph("Ick", [node1, node2, node3], [edge1, edge2, edge3])

Attention, tous les noeuds doivent avoir des données de même type.

STSP.Node — Type

Type représentant les noeuds d'un graphe.

Exemple:

    noeud = Node("James", [π, exp(1)])
    noeud = Node("Kirk", "guitar")
    noeud = Node("Lars", 2)

STSP.PriorityItem — Type

Type représentant un élément de priorité (mutable).

STSP.PriorityQueue — Type

File de priorité.

STSP.Queue — Type

Type représentant une file avec des éléments de type T.

Base.:== — Method

Surcharge de l'égalité

Base.:== — Method

Surcharge de l'égalité

Base.in — Method

Dit si une arête appartient à un ensemble d'arêtes

Base.isless — Method

Définition de la méthode de comparaison pour PriorityItem.

Base.length — Method

Donne le nombre d'éléments sur la file.

Base.popfirst! — Method

Retire et renvoie l'élément ayant la plus basse priorité (poids minimum).

Base.popfirst! — Method

Retire et renvoie l'objet du début de la file.

Base.push! — Method

Ajoute item à la fin de la file s.

Base.show — Method

Affiche un graphe

Base.show — Method

Affiche un arête.

Base.show — Method

Affiche un noeud.

Base.show — Method

Affiche une file.

STSP.Algorithme_HK — Method

Algorithme_HK(graph_edges::Vector{Vector{Int64}}, edge_weights_dict::Dict{Tuple{Int64, Int64}, Float64}, racine::Int64, algo_Arbre_minimal::Int64)

Implémente l'algorithme de Keld Helsgaun.

Arguments

graph_edges::Vector{Vector{Int64}}: Vecteur représentant les arêtes dans le graphe.

edge_weights_dict::Dict{Tuple{Int64, Int64}, Float64}: Dictionnaire stockant les poids des arêtes du graphe.

racine::Int64: Le nœud de départ.

algo_Arbre_minimal: Un entier (1:Kruskal, 2:Prim) qui indique la méthode pour trouver l'arbre de recouvrement minimal d'un graphe.

Retourne

Une liste contenant la tournée minimal du graphe du départ

STSP.Algorithme_Prim — Method

Algorithme_Prim(graph_edges::Vector{Vector{Int64}}, edge_weights_dict::Dict{Tuple{Int64, Int64}, Float64}, start_node::Int64)

Implémente l'algorithme de Prim pour trouver l'arbre de recouvrement minimal d'un graphe.

Arguments

graph_edges::Vector{Vector{Int64}}: Vecteur représentant les arêtes dans le graphe.

edge_weights_dict::Dict{Tuple{Int64, Int64}, Float64}: Dictionnaire stockant les poids des arêtes du graphe.

start_node::Int64: Le nœud de départ de l'algorithme de Prim.

Retourne

Un tuple contenant l'arbre de recouvrement minimal et son poids total.

STSP.Algortihme_Kruskal — Method

(graph_edges::Vector{Vector{Int64}}, edge_weights_dict::Dict{Tuple{Int64, Int64}, Float64})

...

Arguments

graph_edges::Vector{Vector{Int64}}: Vecteur représentant les arêtes dans le graphe considéré

edge_weights_dict::Dict{Tuple{Int64, Int64}, Float64}: Dictionnaire stockant les poids des arêtes du graphe considéré

...

applique l'algorithme de Kruskal afin de renvoyer l'arbre de recouvrement minimal et son poids.

STSP.add_edge! — Method

Ajoute un arete au graphe.

STSP.add_node! — Method

Ajoute un noeud au graphe.

STSP.add_symmetry! — Method

Arguments

a::Dict{Tuple{Int64, Int64}, Float64}: Dictionnaire représentant les poids

Modifie

a::Dict{Tuple{Int64, Int64}, Float64} contenant l'ensemble des poids dans le graphe en prenant en compte les matrices adjacentes supérieures ou inférieures

STSP.affichage_RSL — Method

(filename::String)

...

Arguments

filename::String: Chemin du fichier au format .tsp stockant le graphe considéré

...

Affiche l'application de l'algorithme de Kruskal en rouge sur le graphe considéré.

STSP.affichage_arbre_minimal_kruskal — Method

(filename::String)

...

Arguments

filename::String: Chemin du fichier au format .tsp stockant le graphe considéré

...

Affiche l'application de l'algorithme de Kruskal en rouge sur le graphe considéré.

STSP.affichage_arbre_minimal_prim — Method

(filename::String)

...

Arguments

filename::String: Chemin du fichier au format .tsp stockant le graphe considéré

...

Affiche l'application de l'algorithme de Kruskal en rouge sur le graphe considéré.

STSP.complete_graph_edges — Method

Arguments

graph_edges::Vector{Vector{Int64}}: Vecteur représentant les arêtes dans le graphe.

Retourne

graph_edges::Vector{Vector{Int64}} contenant l'ensemble des arêtes dans le graphe en prenant en compte les matrices adjacentes supérieures ou inférieures

STSP.data — Method

Renvoie le poids de l'arete.

STSP.data — Method

Renvoie les données contenues dans le noeud.

STSP.decalage — Method

decalage(removed_node_dict, removed_edge_vec, removed_weights_dict, racine)

Réalise un décalage d'indice pour trouver un sous arbre de recouvrement minimal

Arguments

removed_node_dict: Dictionnaire contenant le sous graphe sans le noeud racine.

removed_edge_vec: Vecteur contenant les arêtes sans les arêtes avec le noeud racine.

removed_weights_dict: Dictionnaire contenant les poids des arêtes du sous graphe.

racine::Int64: Le nœud de départ.

Retourne

removed_node_dict_copy : Dictionnaire du sous graphe mais avec les indices décalées de 1.

removed_edge_vec_copy : Vecteur du sous graphe mais avec les indices décalées de 1.

removed_weights_dict_copy : Dictionanire du sous graphe mais avec les indices des arêtes décalées de 1.

STSP.edges — Method

Renvoie la liste des aretes du graphe.

STSP.get_degree — Method

Renvoie le degré d'un noeud

STSP.get_one_tree — Method

get_one_tree(graph_edges::Vector{Vector{Int64}}, edge_weights_dict::Dict{Tuple{Int64, Int64}, Float64}, racine::Int64, algo_Arbre_minimal::Int64)

Implémente la méthode pour trouver un 1-tree minimum avec la racine racine

Arguments

graph_edges::Vector{Vector{Int64}}: Vecteur représentant les arêtes dans le graphe.

edge_weights_dict::Dict{Tuple{Int64, Int64}, Float64}: Dictionnaire stockant les poids des arêtes du graphe.

racine::Int64: Le nœud de départ.

algo_Arbre_minimal: Un entier (1:Kruskal, 2:Prim) qui indique la méthode pour trouver l'arbre de recouvrement minimal d'un graphe.

Retourne

Un 1-tree minimum avec la racine racine

STSP.increment_edges — Method

Décalage pour les arêtes

STSP.increment_nodes — Method

Décalage pour les noeuds

STSP.is_empty — Method

Indique si la file est vide.

STSP.n_nodes_to_read — Method

Fonction auxiliaire de read_edges, qui détermine le nombre de noeud à lire en fonction de la structure du graphe.

STSP.name — Method

Renvoie le nom de l'arete.

STSP.name — Method

Renvoie le nom du graphe.

STSP.name — Method

Renvoie le nom du noeud.

STSP.nb_edges — Method

Renvoie le nombre de aretes du graphe.

STSP.nb_nodes — Method

Renvoie le nombre de noeuds du graphe.

STSP.nodes — Method

Renvoie la liste des noeuds du graphe.

STSP.noeud_1 — Method

Renvoie le noeud 1.

STSP.noeud_2 — Method

Renvoie le noeud 2.

STSP.one_tree! — Method

Construction d'un 1-tree à partir d'un arbre de recouvrement

STSP.parcours_preordre — Method

parcours_preordre(tree::Dict{Int64, Vector{Int64}}, node::Int64, visited::Vector{Int64})

Arguments

tree::Dict{Int64, Vector{Int64}}: Un dictionnaire représentant l'arbre couvrant minimal. Chaque clé dans le dictionnaire est un nœud, et chaque valeur associée est un vecteur contenant les voisins connectés à ce nœud.

node::Int64::Dict{Tuple{Int64, Int64}, Float64}: Le nœud de départ pour le parcours en préordre.

visited::Vector{Int64}: Un vecteur contenant les nœuds déjà visités dans l'ordre où ils ont été parcourus.

Modifie

visited::Vector{Int64}: Ce vecteur est mis à jour à chaque appel récursif et contient l'ordre final des nœuds pour la tournée en préordre.

STSP.plot_graph — Method

Affiche un graphe étant données un ensemble de noeuds et d'arêtes.

Exemple :

graph_nodes, graph_edges = read_stsp("bayg29.tsp")
plot_graph(graph_nodes, graph_edges)
savefig("bayg29.pdf")

STSP.plot_graph — Method

Fonction de commodité qui lit un fichier stsp et trace le graphe.

STSP.read_edges — Method

Analyse un fichier .tsp et renvoie l'ensemble des arêtes sous la forme d'un tableau.

STSP.read_header — Method

Analyse un fichier .tsp et renvoie un dictionnaire avec les données de l'entête.

STSP.read_nodes — Method

Analyse un fichier .tsp et renvoie un dictionnaire des noeuds sous la forme {id => [x,y]}. Si les coordonnées ne sont pas données, un dictionnaire vide est renvoyé. Le nombre de noeuds est dans header["DIMENSION"].

STSP.read_stsp — Method

Renvoie les noeuds et les arêtes du graphe.

STSP.remove_edge! — Method

Supprime une arête du graphe

STSP.remove_node! — Method

Supprime un noeud du graphe

STSP.symetric — Method

Dit si deux arêtes sont symétriques